Remove management of client daemons in ldap class
This will allow the user to set the data in hiera, under a profile
namespace perhaps, and then retrieve those values and pass them into the
multiple parts of the module that may consume that data.  In this case,
the base class and the client daemons.

Without this change, users are expected to duplicate the data on a per
class basis and we assume too much about the target deployment scenario.